Fix a compilation warning.
git-svn-id: http://svn.python.org/projects/python/branches/p3yk@55959 6015fed2-1504-0410-9fe1-9d1591cc4771

+++ Modules/_testcapimodule.c
@@ -728,7 +728,7 @@ void print_delta(int test, struct timeval *s, struct timeval *e)
 		e->tv_sec -=1;
 		e->tv_usec += 1000000;
 	}
-	printf("Test %d: %d.%06ds\n", test, (int)e->tv_sec, e->tv_usec);
+	printf("Test %d: %d.%06ds\n", test, (int)e->tv_sec, (int)e->tv_usec);
 }
 
 static PyObject *
